idempotency: support compacted topics
Idempotency already support the compacted topics; it was turned off by
default because the support isn't complete and may cause premature session
termination but it doesn't lead to a violation so it's ok to turn it on.

Premature session termination?

Idempotency relyes on the applied batches, uses the batch's metadata to fill
its seq table (a map from producer is to the last send record number). When
a batch is overwritten during the compaction we may lose the meta info too
and the seq table may be incomplete.

It leads to situation when a producer relays on the local knowledge of what
was replicated, tries replicating a record with the next seq number and it is
got rejected because the seq table isn't complete (another producer issued a
record with the same key and it overwritten the first during compaction).

The complete support requires Raft snapshots that aren't bound to truncation.

(cherry picked from commit 56fb96c65f2fdc85f637273b7235dc700cd9089b)