cloud_storage: delete cache obsolete directories
cloud_storage cache items are stored in directories with randomly
generated names, so almost for every cache item there's a unique nested
directory. Before this PR this nested directories were not removed, now
they are removed every time when cache item is removed.

The removal if the directory may happen concurrently with creating file
in this directory, in this case the directory will be recreated and
creation of file will be re-tried.

Fixes #3449