transactions: fix false reject
Redpanda used to return invalid_txn_state when a commit detected
an already committed transaction

  1. a client commits a transaction
  2. the transaction is committed
  3. tx coordinator dies just before ack'ing a tx
  4. the client retries the commit
  5. the commit hits server & sees that the transaction may
     be committed

When Redpanda returns invalid_txn_state the client interpreters it
as a reject while the tx is committed. it seems that the server
should roll the tx forward since commit is an idempotent op. but
there is a problem with this solution too:

 1. a client commits a transaction
 2. the transaction is committed
 3. the client initiates next transaction
 4. tx coordinator reboots before the prepare phase
 5. the client retries the commit, tx coordinator sees prepared
    state of the previous transaction
 6. if we roll it forward the client will think that the current
    transaction is committed

so we fail the request with unknow error and let a user to recover

(cherry picked from commit e11114e795fd1f23271af104fec4a5493c5bb712)