cloud_storage: Add 'find' method to manifest
The method can be used to find segment metadata by its offset.
Internaly, the method uses lower_bound to find the candidate and then
checks if the offset matches. It's not possible to have entries with the
same offset and different term so the conflicts are not possible.

(cherry picked from commit e163a5affed7ece2806f7aa531a4183083452540)