docker: build Java packages before other steps
The Java build is very slow -- when it's after
other steps, any changes in those other steps
leads to a rebuild of the Java stuff too.

Mitigate this by bumping the Java packages
to be the first things we build.  There are
probably ways to improve the actual build
time of them too, but this helps.