compat/kstreams: split drivers from examples
The KStreams example classes use the same names as the KStreams tests.
Rename them and distinguish between an example and a driver. This will
improve code readability.

Within this commit are changes for #3032. The main problem in #3032 is
that the internal Java application sometimes fails to count all input
messages. This may be related to the 1min window used within the
program. Re-enable with ok_to_fail for now until a fix is proposed.

This commit also re-enables KStreams wikipedia test since it was
disabled due to #2889. Running the test multiple times (50+) did not
reproduce the issue in #2889. Redpanda has changed alot since #2889
was last seen, so re-enable with ok_to_fail to see if the problem
still exists.