compat/kstreams: split drivers from examples
The KStreams example classes use the same names as the KStreams tests.
Rename them and distinguish between an example and a driver. This will
improve code readability.

Within this commit are changes for #3032. The issue in #3032 is that the
RpKConsumer failed to fetch messages in the AnomalyDetection test.
This commit includes changes to the consume loop, uses the
format option in RpkConsumer, and changes the valid msg condition. At
the least, the system will log more info if the failure occurs again.

This commit also re-enables KStreams wikipedia test since it was
disabled due to #2889. Running the test multiple times (50+) did not
reproduce the issue in #2889. Redpanda has changed alot since #2889
was last seen, so re-enable KStreams Wikipedia to see if the problem
still exists.