rpk: only probe broker IDs when needed
The code for loading the ID->URL mapping was
already present in the on-demand path in
brokerIDToURL -- it was not necessary to
pre-populate this in the constructor.

This speeds things up when instantiating an
admin API client which won't use sendToLeader,
as is the case for many RPK CLI cases.