rpk: selectively skip retry on simple GETs to admin API
These are cases where we never expect a 503.  On connection
errors we will fail out fast instead of repeatedly retrying.

This is especially helpful for mapBrokerIDsToURLs, which would
otherwise have a long runtime if used on a cluster with one
down node, while the request for that one node went through
retry cycles.