cloud_storage: Switch to bree_map in remote_partition
The main data structure in the remote_partition is changed from std::map
to absl::btree_map. This ought to minimize memory fragmentation. The
side effect of the switch is that the iterators are not longer stable.
To fix this problem the special iterator type was added. This iterator
caches the key of the object that it points to. On every access it
fetches the value from the underlying btree_map. The complexity of the
operations increases logarithmicaly. So the increment is no longer O(1)
but O(log N). But it's not a problem sicne the base of the logarithm in
case of btree_map is large and the operations are not frequent. We're
only increment an iterator when we need to go from one segment to
another so the amount of work is tiny compared to segment scanning.

(cherry picked from commit 50a91722cd6518abeba73b07aeea718550c1c16a)