admin: make user create/delete requests idempotent
There are cases during leadership changes where
a client might receive a timeout error even if
the operation did actually write successfully, and
then the client would receive a user_exists error
if they retry.

Implement safe idempotency: the create is only accepted
as an idempotent repeat if the existing records credentials
match the request, not just if the usernames are the same.