config: add bounded_property class
This is a generalization of clamped_property, that
provides a validator as well as clamping the value
to safe bounds.

The clamping is redundant as long as validation
was properly applied before storing value, but
is useful if an invalid value has made it into
the system somehow, for example while upgrading
from an earlier version that had different validity
bounds on the property, or if loading a redpanda.yaml
while central config is disabled.