admin: apply config property validators during patch
Previously we were only detecting exceptions from
YAML parsing, or YAML values that did not convert
cleanly to the value type of the property.

Now we will apply any validator hook that was declared
in the property (unless `force` is true, in which case
invalid values will be accepted and we rely on
properties to clamp input to safe values in set_value).