config/binding: add debug-mode shard-locality checks
Same `oncore` helper as in iobuf: make sure that we
aren't accidentally constructing the binding on
a different shard than it is used in.  This is a
relatively easy mistake to make with bindings, because
one might call `bind()` very early in application
startup and then pass the result into a sharded
service start call.