c/dissemination_service: arm dissemination timer unconditionally
Recent changes in metadata_dissemination service introduced an issue
that lead to situation in which leadership metadata were not
disseminated from the node that started with empty broker set. The issue
was caused by the fact that the timer was armed only when `start()`
method dispatched metadata update request. If the `start` method
returned early timer wasn't armed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>