cluster: fix metrics_reporter for single node clusters
It is intentional that we only start reporting metrics
after at least two raft0 batches (either N raft_configuration
messages for a multi-node cluster, or 1 raft_configuration
plus first topic creation for a single node cluster).

However, we were filtering on raft_configuration as
the batch type, so this was only working reliably on the multi
node case.