test: help pyright deduce the correct return type
Execution of the retry loop really should never fall off the end of the
function. But it wasn't clear how to convinence pyright that that was
the case (including asserting the loop does run _and_ catching generic
exception). So here we brute force the decision by explicitly raising if
we fall out of the loop.

The changes the deduced type from Optional[str] -> str.

Signed-off-by: Noah Watkins <noah@redpanda.com>