kafka/tests: use dynamic revision_id
These tests previously hardcoded revision ID 2 when
pre-populating storage contents for new partitions.  This
was based on assumptions about contents of the controller
log, which are no longer valid when feature manager and
config manager may perform writes early in the lifetime
of a new cluster.

Replace hardcoded revision IDs with a dynamic read of
the controller log offset.