rptest/wasm: Refactor complex test dependency
- Having WasmIdentityTest be a base class for all types of identity
tests made choosing a different log_allow_list impossible.

- Also wasm_test was coupled too deeply to these identity tests, for one
a configurable output rate was not allowed. It was always assumed that
only identity coprocessors, i.e. jobs with a 1:1 transformation rate,
were deployed. This means other tests such as the filter test, could not
leverage any of the other infrastructure code setup in WasmIdentityTest.

- The CHAOS_LOG_ALLOW_LIST should only be in the type of wasm tests that
introduce a failure, the inheritence hierarchy meant that this could not happen.

- The API for determining the rate of records on input/output topics and
even how many coprocessors to deploy was non-obvious, complex, and made
the core code in wasm_test.py hard to read and understand.

- This change simplifies the API to be more explicit, therefore all of
the bits in wasm_test become easier to interpret as they are operating
on explicit types and sets, i.e. no more mixed bag lists of tuples.