test: simplify rpk describe topic return type
The type had effectively been list[Partition] | None. The first source
of the optional aspect was that if no topics existed None was returned.
This was changed to raise an exception. All existing users did not check
for this condition and immediately began iterating over the return
value. One user checked for None and raised an exception manually.

The second source was the return filter() whose condition seemed to
confused the type checker. By using the built-in identity function it
could see through the predicate.

The resulting return type is filter[Partition] without the optional
aspect which fixes all of the type checker errors for users that
immediately iterate over the return value.

Signed-off-by: Noah Watkins <noah@redpanda.com>