admin: add HTTP Basic authentication
The request_authenticator is owned by admin_server,
and acts as a factory for result objects which
contain the authentication result of each request
passed in to authenticate().

admin_server is responsible for calling authenticate()
in the right places.

Basic auth is a generic HTTP function that could be
moved into a shared location in the future: for the moment
it's only used in the admin API, so it's defined here.