config: move example() to type-aware property<>
...rather than base_proprty, so that it can generate
guessed example strings from its type, for properties
that do not specify an explicit example.

This is useful for avoiding writing out a bunch of
boilerplate examples for obvious types like bool.  Even
a seemingly obvious thing like a boolean property
being 'true' for example is useful when generating
helper text for users on validation errors.