r/consensus: flush log when refreshing commited index
Redpanda raft implementation never explicitly flush underlying log when
replicating with relaxed consistency. Record batches visibility is
controlled by high watermark offset which is updated even if log isn't
flushed on majority of nodes. Raft committed index semantics is more
strict. Committed index may only be updated if majority of nodes flushed
given offset to the disk.

State machines which underlies transactions and cloud storage require
snapshots to be persisted before allowing log segments to be evicted and
removed by cleanup policy. Snapshots are only taken when all batches up
to the snapshot offsets were applied to the state machine. State machine
batches are applied up to raft committed offset. When committed index is
not updated no state is applied to state machine.

When any of the persisted state machine snapshot creation is pending it
prevents log segments from being evicted.

All together not updating committed index in relaxed consistency mode
and waiting for snapshots to be taken prevents redpanda from cleaning
old data.

The mechanism we introduced previously to prevent described behavior is
based on the eviction_stm notifying raft about a need of commit index
advancement via `refresh_commit_index` method.

Problem that the implementation of that method had was the fact that it
was forcing log to flush only on the leader node, not the followers.
This is not enough since commit index is update when majority of nodes
reached the same flushed offset. Changed the implementation to flush
log on the followers allowing leader to update committed offset of raft
group.

Signed-off-by: Michal Maslanka <michal@vectorized.io>