security: make `superusers` config live updatable
This involves an interface change to authorizer to
have it self-sync with a config binding rather
than receive add_superuser calls.

The authorizer constructor must take a callback
to generate a binding rather than the binding, because
bindings need to be constructed on the same shard as
the instance will live on.