archival: Fix shutdown deadlock
The following situation triggers a deadlock during shutdown.
- The node is being stopped and scheduler_serivce::stop is invoked.
- The method calls ntp_archiver::stop for every archiver that it has.
- If the archiver operation hangs the shutdown can't proceed and because
  the scheduler_service is stopped before most other services the node
  stays online.

The ntp_archiver::stop can hang if after the manifest update we're
trying to update the archival_meta_stm but all other replica nodes of
the same ntp are already down. This can happen during cluster shutdown
or during normral operation (if some nodes are not available for
whatever reasons).

The code inside the ntp_arichiver sets the timeout via the
retry_chain_node that it passes to archival_meta_stm. But the problem is
that the call to _raft->replicate inside the archival_meta_stm doesn't
allow to set a timeout. So when all replicas are not responsive it will
just wait for the quorum. When this happens during node shutdown we can
expect that the raft group will be stopped and the 'replicate' call will
throw an exception or return an error. But it doesn't happen because the
archival subsytem is being shut down in the begining. And when it hangs
it prevents the partition_manager and other components to be stopped. So
the raft_group never gets stopped.

This commit fixes that by using a ss::with_timeout wrapper. In case of
timeout the exception will be thrown inside the upload loop and the
whole operation will be retried some time later. The ntp_archiver itself
won't be used before that moment.