tests: update RESTART_LOG_ALLOW_LIST for tx edge case
If transactions are enabled and writing to a raft
group while it's shutting down, we get this message.

Like some others on our allow list, this is something
that ultimately it would be nice to clear up at the source,
and the allow list acts as a record of which log messages
are noisy.