tools/cmake_test: allow binary to be non-absolute path
The `--binary` argument required an absolute path to test binaries,
which is inconvenient for interactive use.

Now you may specify just the name of the binary, or a *relative* path to
the binary, and the script will find a path for you, prefering debug
builds.

Absolute paths are still supported as before.