rpk: add `cluster config lint` command
This is to help with upgrades from pre-2022.1.1
to later versions, where the old redpanda.yaml may
contain vestigial cluster config settings that are
migrated to the new central cluster config store.

This command doesn't perfectly preserve whitespace
and ordering, but it does preserve comments, and
ordering of the elements within each section.