c/partitions_leaders_table: expose both leader_id and related term
Exposing leader id and related term if it is available. Previously only
the term related with latest leader_id update was held in partition
leaders table. Currently we store last stable leader term (term of a
leader which was actually elected, in raft protocol term is being
increased every time election is taking place). The term is exposed from
`partition_leaders_table` so caller can use it to establish the term and
fence stale requests.

Signed-off-by: Michal Maslanka <michal@vectorized.io>