redpanda/tests: wait for raft0 init at start
Tests use wait_for_controller_leadership as a
utility to wait for all the cluster setup to
be done before proceeding.

New infrastructure in config_manager and feature_manager
leads to some async execution of config writes to the
controller log in the background, which confuses
some tests that expect all controller writes
to be done before they start.

Extend wait_for_controller_leadership to wait
for these writes to raft0 before proceeding.