rpk: reduce RedpandaConfig to just node properties
The future of this class is to be only used when
modifying redpanda.yaml files, which are now only
for node configuration properties.

Cluster configuration changes will flow through
the new `rpk cluster config ...` commands.

For the moment, leave code that expected DeveloperMode
in place, setting the property via Other['developer_mode']
instead.