cluster: make config_frontend take node_id once at construction
Unit tests may expected this: where they construct multiple nodes,
the global configuration objects do not reflect the correct
node ID after construction is done.  We have to take the node ID
during construction and hold on to it.

It would be nice to clean up some of these unit tests to be a bit
less sensitive to this kind of thing, in future.