config: deprecate enable_admin_api
The admin API is an increasingly important part
of operating a redpanda cluster: if it is disabled
then many rpk commands will not work.  Also, disabling
it creates a chicken & egg problem, because once the admin
API is turned off, there is no way to send a config API
request to turn it back on.

It is still possible to disable the admin API
by setting `admin: []` at the top level of the node configuration
file redpanda.yaml.