md_dissemination: added heuristic to force leader updates
When partition is deleted and then created again its NTP does not change
but all underlying structures are reset. Raft `term` value is
initialized to 0.

In partitions leaders table we use `term` to fence stale metadata
updates. Currently `partitions_leader_table` is not aware of partition
revision. When term is reset after partition is deleted and then created
again all the updates coming from the new partition instance are ignored
since their `term` value is smaller from the one that is currently
stored in `partitions_leaders_table`.

When deleting topic `partitions_leader_table` entries related with its
partition are deleted, however this action is not coordinated on all of
the nodes and this may lead to situation in which metadata dissemination
update will be lost and/or leadership update coming from the partition
that was deleted is applied after the entry was deleted from leaders
table.

Implemented a heuristic that checks if majority of current replicas
agree on the leader value and forcing an update in leader table. This
way even when topic was recreated we eventually force update its
leadership information.

This solution is temporary. We need to add revision id to all RPC
requests related with leadership updates, this way we will know exactly
the partition instance that the update is related with.

Signed-off-by: Michal Maslanka <michal@vectorized.io>