rm_stm: make idempotency less depend on time
Redpanda depends on server time to decide if it's safe to remove
idempotency sessions from the seq table cache. When the server
time is misconfigured it may lead to the availability issues: a
session is being removed while it's still active.

Adding a seq_table_min_size config option to set a threshold of
minimum active sessions which prevents them from being removed
even if they're expired.