config: clean up old-style property constructor
Properties didn't used to have a metadata object:
they just took a `required` flag.  When the metadata
was introduced, the old-style constructor was kept
around to reduce the up-front churn.

These days new properties should all have metadata,
so to avoid confusion, remove the old constructor
and convert remaining uses to the new-style constructor.

For the node config properties, that just means empty
metadata, because at the moment they aren't touched
by an API that would make use of it.