c/health_monitor: added revision_id to partition_status
Added `model::revision_id` field to `cluster::partition_status`
structure. The revision is is used to determine an actual version of
ntp. The revision is important to recognize an instance of the ntp if it
was deleted and then created again.

Note that the version of `partition_status` field was decremented (moved
from 0 to -1) this is a workaround allowing us to maintain backward
compatibility with previous versions of redpanda. If the version field
was incremented old redpanda nodes would assert and restart.

Signed-off-by: Michal Maslanka <michal@vectorized.io>