tests: ignore <1kb memory leaks in ducktape
On investigation, these appear to be Seastar internal structures,
with a relatively rare race during shutdown leading to a leak.

It would be better to get to the root of this, but for the moment
it's better not to have intermittent test failures for something
that does not appear to be a real world bug.