archival: correctly calculate offset gaps metric
Previously, offset gaps metric was incremented in the archival_policy
methods, which could be incorrect because archival_policy can be called
many times without actually uploading any segments. Also the code missed
the case when there is a gap between local segments. We increment this
metric after actually uploading a segment, which is more robust.