config: revert developer_mode to a node property
If this was a cluster configuration property, it became possible
to break a "developer mode" cluster by setting it to false, then
at next startup if the --memory value was below the safety limit,
then redpanda wouldn't start at all.

In general this kind of move is not permitted, but
central config is not yet in the wild, so we can sneak
this in before 22.1.1.

In principle, it made sense for a "developer mode" flag
(controlling some general sense of whether to be permissive)
to be a cluster property.  However, the specific effect
of this property is to permit or forbid certain CLI flag
values for --memory, so it makes sense to take the developer_mode
setting as close to possible to where we set CLI flags (i.e.
on the node itself).