(split) k8s/cluster: Improve memory configuration
In #1416 some changes were made, but I don't think it was intended
to remove `--memory` flag in `developerMode`.

This PR changes the behaviour:
* When `requests.memory` or `redpanda.memory` is set, `--memory` is
  passed regardless of `developerMode`
* When `--memory` is set, also set `--reserve-memory=0M`. There is
  already a 10% buffer calculated in `requests.RedpandaMemory()`

Setting both `--memory` and `--reserve-memory` is undocumented, and
shouldn't be used. However, there is insufficient information to
correctly calculate what `--reserve-memory` should be. By setting
just `--memory`, seastar will reserve at least 1.5Gi for the OS,
which doesn't make sense in a container.