tests: fix running test_secret_redaction locally
This test relied on debug/trace level logs to work,
because part of it's job is to test we aren't leaking
secrets to those logs.

That always worked in CI and on my local setup where
I set redpanda log level to trace, but if someone
runs ducktape without any REDPANDA_LOG_LEVEL argument,
it will fail.

Fix by explicitly setting the redpanda log level for this
test.