r/consensus: added api allowing aborting ongoing configuration change
Added API that allows caller to abort ongoing configuration change.

Two APIs were added

`revert_configuration_change` - this api is safe and will never cause
the data loss. The API changes the direction of configuration migration
in joint consensus. i.e. desired configuration becomes an old one and
the old one becomes desired. This approach will allow raft to handle all
necessary recoveries even if some of the nodes were already demoted to
learners.

`abort_configuration_change` - this is a method of last resort that may
be required to recover from catastrophic failures. f.e. single replica
was moved between nodes and it is stuck in the middle of move. This
method is not always safe as it may lead to data loss. This should be
used for catastrophic failure recovery.

Signed-off-by: Michal Maslanka <michal@vectorized.io>