config: add features_auto_enable
This property lets people decide whether they want
new features to switch on by default, or to have to
manually activate features later.

The default is auto-activation: it is expected that
this will only be set false in tightly managed environments
where an external management component is going to activate
features after they have determined that an upgrade
was successful and are willing to pass a point of no return.