cloud_storage: Store index files in the cache
When the remote_segment is materialized the following cases are
possible:
- The segment is hydrated;
- The segment is not hydrated;

Currently, the index is built when the segment is being hydrated. This
simplifies the design a lot because the alternative is to build the
index lazily (when we read from the segment). Because there is no
guarantee that the user will be reading the segment sequentially we
can't rely on this approach. So current implementation builds the index
while it's being downloaded from S3.

The problem is that the remote_segment might be evicted. In this case
the segment will remain hydrated in cache but the in-memory index will
be deleted. To solve this problem this commit introduced the
serialization/deserialization logic to offset_index.

The offset_index gets serialized after it was built. The code appends
".index" prefix to the segment name. When the remote_segment receives
hydration request it checks if the segment is present in the cache. If
it's not present in the case the hydration process starts that leads to
the creation of the in-memory offset_index. This index gets serialized.
Alternatively, if the segment is already hydrated the remote_segment
tries to read the serialized offset_index from cache.

The cache should be able to deal with index files just fine. It doesn't
care about file extension unless it's ".part" file.