cluster: implement feature barrier operation
This is a synchronisation operation for a one-shot barrier
that enables data migration code to ensure all nodes have
reached the same point before changing on-disk state.

In the case of consumer offsets, this is for waiting
until all nodes have seen "preparing" state (i.e. stopped
accepting writes) before proceeding with migration.

This barrier is pretty simple and relies on the properties
of callers for correctness: they must be using something
persistent (in this case pending feature state) to prompt their
entry to the barrier, and something else persistent (in this
case active feature state) to prompt them to skip the barrier
in future.  That enables us to cope with nodes restarting
during the barrier simply.