r/consensus: added configuration change in progress metric
Added gauge metric indicating that the raft group configuration is being
changed (i.e. it is in joint state). The metric value of 1 indicates
that the configuration is in joint state. The metric is only reported by
the leader.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 3216043afc65d35125fa0304a14d45ede3aad683)