s/log_manager: handle race condition when deleting topic directory
Topic directory deletion process in redpanda is based on the fibers
dispatched after deleting each partition. Last partition deleted from
the topic will eventually delete the topic directory. The deletion
process may race with the process of creating a topic with the same
name. When this happen the deletion fiber may try to delete a topic
directory which is not empty.

Added handling of the situation in which topic is identified as empty
but can not be deleted as the new partition was already created in the
same directory.

Signed-off-by: Michal Maslanka <michal@vectorized.io>