tests/partition_movement: extend timeout for partition move
Since `__consumer_offsets` topic is now being moved as a part of
partition movement test we may need to wait for the result a little bit
longer. We need to consider that the topic is being used by the
consumers and it is compacted. Considering constrained resources and
debug build it may not manage to be moved in shorted time period making
the tests flaky.

Signed-off-by: Michal Maslanka <michal@vectorized.io>