config: do not report string properties as arrays
In actual fact they _are_ collections, but it's
confusing to propagate that fact up the stack, and
most people dont think of a string as an array
of characters.

Fix by special casing our type code to not consider
common string types as arrays.

This wasn't noticed before because we never had
properties that were simple strings: they were always
vector<> or optional<> strings.