cluster: Add rack constraint to partition allocator
The rack placement contraint is a soft constraint. In the ideal
situation it will put every replica on a node in a different rack.
If we don't have enough racks the replica placement will be affected
by other constraints and some replicaas will end up in the same rack.

In the situation when the rack-id is not set for a node it will assume a
unique rack.