rpk: do not require lib/libexec files
RPK has logic for probing for an installation
file by looking for redpanda binaries.  It
is not just looking for the actual entrypoints,
but also for the libexec files: this is an
implementation detail that rpk should not
care about.

Motivated by https://github.com/redpanda-data/vtools/pull/528 ,
in which rpk stops working when we change rpk to put its
binary in bin/rpk rather than in libexec with a wrapper script.