k/response_writer: use term as a batch partition leader epoch
In redpanda we use term as a source for leader epoch information.
Previously we were not setting partition leader epoch in batch header
and always returned 0. This was not causing any issues as leader epoch
wasn't supported at all in redpanda. Now when we support leader epoch in
all other request we must return correct value in record batch header.

Signed-off-by: Michal Maslanka <michal@vectorized.io>