kafka/s: Tunable request rate option
- Adds the configuration parameter `target_quota_time_rate` which
may be used to enable/disable request rate throttling (default disabled)
or to increase/decrease the threshold dynamically

- Note this is different then how the `request_percentage` parameter
works in kafka today. In Kafka this parameter defines (per
user/client/<user/client>) how much of the quota window is allowed to be
consumed by an entity. Since the quota manager is at the moment not kafka
`user` aware, it was determined this would be the best usage of such a parameter.