k/group: do not create metrics for kafka_internal group manager
Added configuration option to disable all offset metrics for the groups
managed by given `group_manager` instance.

We do not want to expose metrics for groups managed by the old group
manager (the one using `kafka_internal/group` topic). Exposing metrics
for those group is not required as the metrics will be exported by new
manager instance and it may lead to migration errors in which new
manager will not be able to create a group because of metric being
registered twice.

Signed-off-by: Michal Maslanka <michal@vectorized.io>