s/parser: fixed reading batches with header_crc equal to 0
When redpanda is interrupted it may happen that some of the log segment
files will be fallocated but not truncated to valid range. In order to
recognize this situation and skip parsing `storage::parser` used only
`header_crc` field value. This lead to situation in which user could not
read a valid batch if its header crc happen to be equal to 0.

Fixed this issue by checking if the whole iobuf header is filled with
zeros

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 056b711ae8021981ae8a5006f3fd2150f347e720)