admin: improved validation when enabling cloud storage
cloud_storage_enabled is special, in that when it is true, several
other properties must be non-null.  Otherwise Redpanda won't start.

This is a trap for the unwary: make it safe by adding more
validation, so that if someone tries to enable the setting
without the required properties set, that will be rejected.