cluster: Improve rack-aware replica placement
Rack aware placement of replicas is implemented using soft constraint.
The scores which was returned by the constraint were 0 and 1. With
scores like this it was still pretty common for the solver to put all
replicas on the same rack.

This commit changes the behavior by using different scores. The score
is still 0 if two nodes have the same rack id. But if the rack ids are
different the score is 'max_score/num_brokers'. This way the scoring
works in favor of rack placement even if it will breaks some other soft
constraints.