r/mux_state_machine: removed mutex protecting offsets
The `mux_state_machine` contained `mutex` serializing all replicate
calls to underlying raft group. The mutex was required since it
was used to wait for replicate to finish before `apply` was called by
the fiber applying batches to stm state.

The use of mutex may lead to a deadlock. The deadlock was triggered when
following events happened:

1. Raft group recovered and starts applying multiple batches to stm
2. Every batch apply requires `mux_state_machine::_mutex` to be acquired
   Batch applies are executed in `applying_consumer`
3. Replicate call was called via `mux_state_machine::replicate` this
   acquires the `_mutex` and waits for replicate to finish
4. Raft group leadership was lost
5. Group received append_entries_request from new leader that requires
   log truncation

This leads to a deadlock in which `log::truncate` is waiting for `log`
write lock which can not be obtained as log reader is alive. Log reader
consumer is waiting for `mux_state_machine::_mutex` which can not be
obtained as `replicate` call is waiting under the `_mutex` for
replication to finish. Replication can not finish since it stuck in
`log:truncate`

Solved this issue by removing mutex from `mux_state_machine`. Instead of
waiting for offset to be assigned after replicate we keep application
results for all the offsets if any of replicate calls are pending this
way when replicate finally finishes we may lookup application result in
results map.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 552d97c526da8aa4c68b94c8fcf9b45f58add0fd)

 Conflicts:
	src/v/raft/mux_state_machine.h

 The spawn_with_gate utility wasn't backported into v21.11.x so had to
 manually use (void) to throw away the future to background it.