archival: run each ntp_archiver in its own fiber
Each ntp_archiver process now runs in its own fiber, instead of being
externally driven by the archival::service object. This allows for early
exits when the archiver notices that this shard is no longer the leader
for a particular ntp and reduces the chance that there will be several
archivers running concurrently.