ssx: Add this_fiber_id function
Add 'this_fiber_id' function that returns unique id of the fiber.
The implementation is based on seastar::task and uses 'waiting_task' to
figure out dependencies between individual tasks. Normally, a fiber
would be executed using multiple tasks so using address of the task
itself is not an option.