ssx: Add fiber local storage implementation
The same idea as thread local storage. The differences are:
- different interface, the creation of the fiber_local object and access
  are separated.
- it is possible to have very large number of unique futures. also, it's
  also common to have high churn (fibers are created and destroyed all
the time) so it's essential to remove data from fiber local storage once
the fiber is completed.

The fiber_local template is used to create fiber local objects:

  struct my_tag_t{};
  ssx::fiber_local<my_tag_t, int> container(42);

The first type parameter is a tag type. To access the content of the
fiber_local instance the following code might be used:

  ssx::fiber_local_selector<my_tag_t, int> selector;
  assert(selector.get() == 42);

Note that there is no need to pass fiber_local instance to the c-tor of
the selector object. They're connected using thread_local variable. This
thread_local variable is created for every unique combination of the tag
type and value type.

If the fiber_local instance is destoryed before the selector it is
possible to trigger an internal assertion while caling get or set
methods of the selector. In order to avoid this it is possible to check
selector for validity using 'has_value' method or to use 'try_get' and
'try_set' methods.

The latency of the fiber_local access is below 100ns (50-60ns in my
testing) if the number of fibers that created fiber_local instance with
the same parameters is low (less than 1000). It probably might grow
a higher with increased concurrency.

The usage of fiber_local might be an antipattern in many cases (same as
normal TLS). But it has few legit usecases that make it worth to have.