Prevent quadratic time in leader balancer
In the existing leader balancer code, we loop over all shards twice in
a nested loop, which is quadratic in the number of shards.

For large number of shards (say > 1000), this can stall the reactor for
more than a second, which at best may cause reduced performance for
other requests hitting this shard and at worst may cause timeouts which
destabilize the cluster.

This change switches to an algorithm that is linear in the number of
shards but should make the same decisions as the existing one:

We loop over the shards from highest to lowest load and then for each
raft group (parition) led by this group we check if can move the leader
to a lower load shard. We track the best candidate seen so far and
at the end of handling each shard if any candidate has been found which
improves the error we stop the search and return it.

This does not always find the *best* movement (e.g., the next-highest
loaded shard may have a better movement since it leads a group which
has a replica on a less-loaded shard, with the difference enough to
overcome the higher load on this cluster), but it should come close in
most pratical situations and it is the same logic as we are currently
using.

Performance results:

Using 72 nodes x 16 shards x 80 replicas:

version     test              median
original    bench_movement    1.663s
original    bench_all         1.600s
updated     bench_movement   3.393ms
updated     bench_all        6.653ms

Here "orignal" is before this change and "updated" is with this change.

The bench_movement test is the find_movement function, which had the
quadratic behavior, while "all" include also the construction of the
balancer object, which was insigificant before but now takes
about half the time.

Speedup is 490x for these inputs for find_movement, or 250x if you
consider initial construction as well.

I've checked by varying the parameters that performance appears linear
in the number of shards and linear in the number of partitions
per shard. So in the usual case of scaling out a deployment by adding
more shards and keeping the paritions per shard roughly constant
this balancing cost should increase only linearly, giving us a
lot of headroom before performance again becomes a problem.

Fixes #4172.